Add keybinding to insert a line break
This keybinding adds a line break after the currently active line,
if multi-line is enabled.

I had to add `app$session$app$timeoutlen <- 0.2` to make it bearable.
The default value is `1` and I don't have the patience for that. I
can barely notice a little lag with 0.2, but it's tolerable.